Refactor most utils to a single file
