--- src/alt/utils/formatAsConstant.js
@@ -1,6 +1,0 @@
-export default function formatAsConstant(name) {
-  return name.replace(/[a-z]([A-Z])/g, (i) => {
-    return `${i[0]}_${i[1].toLowerCase()}`
-  }).toUpperCase()
-}
-
