--- src/alt/utils/getInternalMethods.js
@@ -1,17 +1,0 @@
-/* istanbul ignore next */
-function NoopClass() { }
-
-const builtIns = Object.getOwnPropertyNames(NoopClass)
-const builtInProto = Object.getOwnPropertyNames(NoopClass.prototype)
-
-export default function getInternalMethods(obj, isProto) {
-  const excluded = isProto ? builtInProto : builtIns
-  return Object.getOwnPropertyNames(obj).reduce((value, m) => {
-    if (excluded.indexOf(m) !== -1) {
-      return value
-    }
-
-    value[m] = obj[m]
-    return value
-  }, {})
-}
