--- src/alt/utils/uid.js
@@ -1,8 +1,0 @@
-export default function uid(container, name) {
-  let count = 0
-  let key = name
-  while (Object.hasOwnProperty.call(container, key)) {
-    key = name + String(++count)
-  }
-  return key
-}
