--- src/alt/utils/warnings.js
@@ -1,6 +1,0 @@
-export function warn(msg) {
-  /* istanbul ignore else */
-  if (typeof console !== 'undefined') {
-    console.warn(new ReferenceError(msg))
-  }
-}
